requirements.txt
PyQt6
pyqtdarktheme
opencv-python
pyserial
numpy
